refactor(elevenlabs): ModelCatalog replaced by OllamaApiCatalog
