+++ src/ai-bundle/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
  * The `api_catalog` option for `Ollama` has been removed as the catalog is now automatically fetched from the Ollama server
  * The `api_key` option for `Ollama` is now `null` by default to allow the usage of a `ScopingHttpClient`
  * The `endpoint` option for `Ollama` is now `null` by default to allow the usage of a `ScopingHttpClient`
+ * The `api_catalog` option for `ElevenLabs` has been removed as the catalog is now automatically fetched from the ElevenLabs servers
+ * The `api_key` option for `ElevenLabs` is now `null` by default to allow the usage of a `ScopingHttpClient`
 
 0.6
 ---
