+++ src/ai-bundle/tests/DependencyInjection/AiBundleTest.php
@@ -34,7 +34,6 @@ use Symfony\AI\Chat\ManagedStoreInterface as ManagedMessageStoreInterface;
 use Symfony\AI\Chat\MessageStoreInterface;
 use Symfony\AI\Platform\Bridge\Cache\CachePlatform;
 use Symfony\AI\Platform\Bridge\Decart\PlatformFactory as DecartPlatformFactory;
-use Symfony\AI\Platform\Bridge\ElevenLabs\ModelCatalog as ElevenLabsModelCatalog;
 use Symfony\AI\Platform\Bridge\ElevenLabs\PlatformFactory as ElevenLabsPlatformFactory;
 use Symfony\AI\Platform\Bridge\Failover\FailoverPlatform;
 use Symfony\AI\Platform\Bridge\Failover\FailoverPlatformFactory;
@@ -45,7 +44,6 @@ use Symfony\AI\Platform\Message\TemplateRenderer\ExpressionLanguageTemplateRende
 use Symfony\AI\Platform\Message\TemplateRenderer\StringTemplateRenderer;
 use Symfony\AI\Platform\Message\TemplateRenderer\TemplateRendererRegistry;
 use Symfony\AI\Platform\Model;
-use Symfony\AI\Platform\ModelCatalog\ModelCatalogInterface;
 use Symfony\AI\Platform\PlatformInterface;
 use Symfony\AI\Store\Bridge\AzureSearch\SearchStore as AzureStore;
 use Symfony\AI\Store\Bridge\Cache\Store as CacheStore;
@@ -4140,18 +4138,15 @@ class AiBundleTest extends TestCase
         $this->assertSame([ElevenLabsPlatformFactory::class, 'create'], $definition->getFactory());
         $this->assertTrue($definition->isLazy());
 
-        $this->assertCount(7, $definition->getArguments());
+        $this->assertCount(5, $definition->getArguments());
         $this->assertSame('https://api.elevenlabs.io/v1/', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(2));
         $this->assertSame('http_client', (string) $definition->getArgument(2));
-        $this->assertFalse($definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(3));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(4));
-        $this->assertSame('ai.platform.model_catalog.elevenlabs', (string) $definition->getArgument(4));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(5));
-        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(5));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(6));
-        $this->assertSame('event_dispatcher', (string) $definition->getArgument(6));
+        $this->assertSame('event_dispatcher', (string) $definition->getArgument(4));
 
         $this->assertTrue($definition->hasTag('proxy'));
         $this->assertSame([['interface' => PlatformInterface::class]], $definition->getTag('proxy'));
@@ -4161,14 +4156,7 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasAlias(PlatformInterface::class.' $elevenlabs'));
         $this->assertTrue($container->hasAlias(PlatformInterface::class));
 
-        $modelCatalogDefinition = $container->getDefinition('ai.platform.model_catalog.elevenlabs');
-
-        $this->assertSame(ElevenLabsModelCatalog::class, $modelCatalogDefinition->getClass());
-        $this->assertTrue($modelCatalogDefinition->isLazy());
-
-        $this->assertTrue($modelCatalogDefinition->hasTag('proxy'));
-        $this->assertSame([['interface' => ModelCatalogInterface::class]], $modelCatalogDefinition->getTag('proxy'));
-
+        // Custom endpoint (v2, etc) + API key
         $container = $this->buildContainer([
             'ai' => [
                 'platform' => [
@@ -4186,18 +4174,15 @@ class AiBundleTest extends TestCase
         $this->assertSame([ElevenLabsPlatformFactory::class, 'create'], $definition->getFactory());
         $this->assertTrue($definition->isLazy());
 
-        $this->assertCount(7, $definition->getArguments());
+        $this->assertCount(5, $definition->getArguments());
         $this->assertSame('https://api.elevenlabs.io/v2/', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(2));
         $this->assertSame('http_client', (string) $definition->getArgument(2));
-        $this->assertFalse($definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(3));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(4));
-        $this->assertSame('ai.platform.model_catalog.elevenlabs', (string) $definition->getArgument(4));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(5));
-        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(5));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(6));
-        $this->assertSame('event_dispatcher', (string) $definition->getArgument(6));
+        $this->assertSame('event_dispatcher', (string) $definition->getArgument(4));
 
         $this->assertTrue($definition->hasTag('proxy'));
         $this->assertSame([['interface' => PlatformInterface::class]], $definition->getTag('proxy'));
@@ -4207,14 +4192,7 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasAlias(PlatformInterface::class.' $elevenlabs'));
         $this->assertTrue($container->hasAlias(PlatformInterface::class));
 
-        $modelCatalogDefinition = $container->getDefinition('ai.platform.model_catalog.elevenlabs');
-
-        $this->assertSame(ElevenLabsModelCatalog::class, $modelCatalogDefinition->getClass());
-        $this->assertTrue($modelCatalogDefinition->isLazy());
-
-        $this->assertTrue($modelCatalogDefinition->hasTag('proxy'));
-        $this->assertSame([['interface' => ModelCatalogInterface::class]], $modelCatalogDefinition->getTag('proxy'));
-
+        // Custom http client + API key
         $container = $this->buildContainer([
             'ai' => [
                 'platform' => [
@@ -4232,18 +4210,15 @@ class AiBundleTest extends TestCase
         $this->assertSame([ElevenLabsPlatformFactory::class, 'create'], $definition->getFactory());
         $this->assertTrue($definition->isLazy());
 
-        $this->assertCount(7, $definition->getArguments());
+        $this->assertCount(5, $definition->getArguments());
         $this->assertSame('https://api.elevenlabs.io/v1/', $definition->getArgument(0));
         $this->assertSame('foo', $definition->getArgument(1));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(2));
         $this->assertSame('foo', (string) $definition->getArgument(2));
-        $this->assertFalse($definition->getArgument(3));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(3));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(4));
-        $this->assertSame('ai.platform.model_catalog.elevenlabs', (string) $definition->getArgument(4));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(5));
-        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(5));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(6));
-        $this->assertSame('event_dispatcher', (string) $definition->getArgument(6));
+        $this->assertSame('event_dispatcher', (string) $definition->getArgument(4));
 
         $this->assertTrue($definition->hasTag('proxy'));
         $this->assertSame([['interface' => PlatformInterface::class]], $definition->getTag('proxy'));
@@ -4253,45 +4228,32 @@ class AiBundleTest extends TestCase
         $this->assertTrue($container->hasAlias(PlatformInterface::class.' $elevenlabs'));
         $this->assertTrue($container->hasAlias(PlatformInterface::class));
 
-        $modelCatalogDefinition = $container->getDefinition('ai.platform.model_catalog.elevenlabs');
-
-        $this->assertSame(ElevenLabsModelCatalog::class, $modelCatalogDefinition->getClass());
-        $this->assertTrue($modelCatalogDefinition->isLazy());
-
-        $this->assertTrue($modelCatalogDefinition->hasTag('proxy'));
-        $this->assertSame([['interface' => ModelCatalogInterface::class]], $modelCatalogDefinition->getTag('proxy'));
-
+        // Scoped http client (already configured with API key)
         $container = $this->buildContainer([
             'ai' => [
                 'platform' => [
                     'elevenlabs' => [
-                        'api_key' => 'foo',
-                        'api_catalog' => true,
+                        'http_client' => 'custom_scoped',
                     ],
                 ],
             ],
         ]);
 
         $this->assertTrue($container->hasDefinition('ai.platform.elevenlabs'));
-        $this->assertTrue($container->hasDefinition('ai.platform.model_catalog.elevenlabs'));
 
         $definition = $container->getDefinition('ai.platform.elevenlabs');
-
-        $this->assertTrue($definition->isLazy());
         $this->assertSame([ElevenLabsPlatformFactory::class, 'create'], $definition->getFactory());
+        $this->assertTrue($definition->isLazy());
 
-        $this->assertCount(7, $definition->getArguments());
+        $this->assertCount(5, $definition->getArguments());
         $this->assertSame('https://api.elevenlabs.io/v1/', $definition->getArgument(0));
-        $this->assertSame('foo', $definition->getArgument(1));
+        $this->assertNull($definition->getArgument(1));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(2));
-        $this->assertSame('http_client', (string) $definition->getArgument(2));
-        $this->assertTrue($definition->getArgument(3));
+        $this->assertSame('custom_scoped', (string) $definition->getArgument(2));
+        $this->assertInstanceOf(Reference::class, $definition->getArgument(3));
+        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(3));
         $this->assertInstanceOf(Reference::class, $definition->getArgument(4));
-        $this->assertSame('ai.platform.model_catalog.elevenlabs', (string) $definition->getArgument(4));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(5));
-        $this->assertSame('ai.platform.contract.elevenlabs', (string) $definition->getArgument(5));
-        $this->assertInstanceOf(Reference::class, $definition->getArgument(6));
-        $this->assertSame('event_dispatcher', (string) $definition->getArgument(6));
+        $this->assertSame('event_dispatcher', (string) $definition->getArgument(4));
 
         $this->assertTrue($definition->hasTag('proxy'));
         $this->assertSame([['interface' => PlatformInterface::class]], $definition->getTag('proxy'));
@@ -4300,14 +4262,6 @@ class AiBundleTest extends TestCase
 
         $this->assertTrue($container->hasAlias(PlatformInterface::class.' $elevenlabs'));
         $this->assertTrue($container->hasAlias(PlatformInterface::class));
-
-        $modelCatalogDefinition = $container->getDefinition('ai.platform.model_catalog.elevenlabs');
-
-        $this->assertSame(ElevenLabsModelCatalog::class, $modelCatalogDefinition->getClass());
-        $this->assertTrue($modelCatalogDefinition->isLazy());
-
-        $this->assertTrue($modelCatalogDefinition->hasTag('proxy'));
-        $this->assertSame([['interface' => ModelCatalogInterface::class]], $modelCatalogDefinition->getTag('proxy'));
     }
 
     public function testFailoverPlatformCanBeCreated()
