+++ src/platform/src/Bridge/ElevenLabs/PlatformFactory.php
@@ -13,7 +13,6 @@ namespace Symfony\AI\Platform\Bridge\ElevenLabs;
 
 use Symfony\AI\Platform\Bridge\ElevenLabs\Contract\ElevenLabsContract;
 use Symfony\AI\Platform\Contract;
-use Symfony\AI\Platform\ModelCatalog\ModelCatalogInterface;
 use Symfony\AI\Platform\Platform;
 use Symfony\Component\HttpClient\EventSourceHttpClient;
 use Symfony\Component\HttpClient\ScopingHttpClient;
@@ -29,8 +28,6 @@ final class PlatformFactory
         string $endpoint = 'https://api.elevenlabs.io/v1/',
         #[\SensitiveParameter] ?string $apiKey = null,
         ?HttpClientInterface $httpClient = null,
-        bool $apiCatalog = false,
-        ModelCatalogInterface $modelCatalog = new ModelCatalog(),
         ?Contract $contract = null,
         ?EventDispatcherInterface $eventDispatcher = null,
     ): Platform {
@@ -47,7 +44,7 @@ final class PlatformFactory
         return new Platform(
             [new ElevenLabsClient($httpClient)],
             [new ElevenLabsResultConverter($httpClient)],
-            $apiCatalog ? new ElevenLabsApiCatalog($httpClient) : $modelCatalog,
+            new ModelCatalog($httpClient),
             $contract ?? ElevenLabsContract::create(),
             $eventDispatcher,
         );
