+++ src/platform/src/Bridge/ElevenLabs/Tests/PlatformFactoryTest.php
@@ -12,7 +12,6 @@
 namespace Symfony\AI\Platform\Bridge\ElevenLabs\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\AI\Platform\Bridge\ElevenLabs\ElevenLabsApiCatalog;
 use Symfony\AI\Platform\Bridge\ElevenLabs\ModelCatalog;
 use Symfony\AI\Platform\Bridge\ElevenLabs\PlatformFactory;
 use Symfony\Component\HttpClient\HttpClient;
@@ -27,13 +26,6 @@ final class PlatformFactoryTest extends TestCase
         $this->assertInstanceOf(ModelCatalog::class, $platform->getModelCatalog());
     }
 
-    public function testStoreCanBeCreatedWithoutScopingHttpClientAndApiCatalog()
-    {
-        $platform = PlatformFactory::create(apiKey: 'foo', httpClient: HttpClient::create(), apiCatalog: true);
-
-        $this->assertInstanceOf(ElevenLabsApiCatalog::class, $platform->getModelCatalog());
-    }
-
     public function testStoreCanBeCreatedWithScopingHttpClient()
     {
         $platform = PlatformFactory::create(httpClient: ScopingHttpClient::forBaseUri(HttpClient::create(), 'https://api.elevenlabs.io/v1/', [
@@ -43,16 +35,5 @@ final class PlatformFactoryTest extends TestCase
         ]));
 
         $this->assertInstanceOf(ModelCatalog::class, $platform->getModelCatalog());
-    }
-
-    public function testStoreCanBeCreatedWithScopingHttpClientAndApiCatalog()
-    {
-        $platform = PlatformFactory::create(httpClient: ScopingHttpClient::forBaseUri(HttpClient::create(), 'https://api.elevenlabs.io/v1/', [
-            'headers' => [
-                'xi-api-key' => 'bar',
-            ],
-        ]), apiCatalog: true);
-
-        $this->assertInstanceOf(ElevenLabsApiCatalog::class, $platform->getModelCatalog());
     }
 }
