+++ src/platform/src/Bridge/ElevenLabs/phpstan.dist.neon
@@ -3,7 +3,7 @@ includes:
     - ../../../../../.phpstan/extension.neon
 
 parameters:
-    level: 6
+    level: 8
     paths:
         - .
         - Tests/
